// app/src/generated/NodeValue.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ScalarForm } from "./ScalarForm";

export type NodeValue = { "kind": "scalar", text: string, form: ScalarForm, } | { "kind": "list", count: number, } | { "kind": "block", count: number, };